feat: Add DEFAULT_AVATAR_URL to .env.example and use it in signUp controller
This commit adds the DEFAULT_AVATAR_URL environment variable to the .env.example file. It also updates the signUp controller in the auth.controller.js file to set the avatar field of the new user to the value of DEFAULT_AVATAR_URL from the environment variables. This change ensures that a default avatar URL is assigned to new users during the signup process.