feat: Add avatar folder deletion when deleting user
This commit modifies the deleteUserAvatar function in the user.utils.js file to delete the user's avatar folder from the cloud storage when deleting the user. It calls the checkAndDeleteFolder function from the cloudinary.js file to handle the deletion. This change ensures that the user's avatar folder is properly cleaned up when the user is deleted, improving data management and storage efficiency.