feat: Add react-lottie dependency for animation support
This commit adds the react-lottie dependency to the frontend package.json file. The react-lottie library provides support for animations in React applications. This dependency will be used to display a check animation in the FinishStep component. By adding react-lottie, we can enhance the user experience and make the room creation/update process more visually appealing.