feat: Add createUser function to users.controller
This commit adds the `createUser` function to the `users.controller.js` file. The function handles the creation of a new user based on the provided request body. It checks if the user is an administrator and assigns the appropriate role accordingly. The user's password is encrypted before saving it to the database. If an error occurs during the creation process, an error message is returned. This addition enhances the functionality of the user management system.