refactor: Populate roles in getUser function
This commit updates the `getUser` function in the `users.controller.js` file to populate the `roles` field of the user object. This change ensures that the roles associated with the user are included in the response. By populating the roles, we provide more comprehensive information about the user and enhance the functionality of the user management system.