refactor: Add getUserContacts and addUserToContactList functions
This commit adds two new functions, `getUserContacts` and `addUserToContactList`, to the users controller in the backend codebase. The `getUserContacts` function retrieves the contacts of a user and includes the last message exchanged with each contact. The `addUserToContactList` function adds a user to the contact list of another user. These additions enhance the functionality of the users controller by providing the ability to manage user contacts.