refactor: Update all anchors to use Links insead
This commit updates the link in the InstagramIntro component of the About page. The previous link was pointing to '/single-v1/5', but it has been removed to improve the user experience. This change ensures that the link in the component is now a regular link without any specific destination. It enhances the usability and consistency of the About page.