chore: Update login and registration forms to redirect to appropriate dashboard page
This commit updates the login and registration forms to redirect users to the appropriate dashboard page based on their role. If the user is an administrator, they will be redirected to the "/dashboard" page. Otherwise, they will be redirected to the "/dashboard/rooms" page. This change improves the user experience by ensuring that users are directed to the correct dashboard based on their role.